Move FlexibeleEnum class to separate file
